Fix type on create cursor surface

+++ libgitg/gitg-platform-support.c
@@ -56,14 +56,14 @@ gitg_platform_support_http_get_finish (GAsyncResult  *result,
 cairo_surface_t *
 gitg_platform_support_create_cursor_surface (GdkDisplay    *display,
                                              GdkCursorType  cursor_type,
-                                             gint          *hot_x,
-                                             gint          *hot_y,
-                                             gint          *width,
-                                             gint          *height)
+                                             gdouble       *hot_x,
+                                             gdouble       *hot_y,
+                                             gdouble       *width,
+                                             gdouble       *height)
 {
 	GdkCursor *cursor;
 	cairo_surface_t *surface;
-	gint w = 0, h = 0;
+	gdouble w = 0, h = 0;
 
 	cursor = gdk_cursor_new_for_display (display, cursor_type);
 	surface = gdk_cursor_get_surface (cursor, hot_x, hot_y);
