+++ libgitg/gitg-platform-support.h
@@ -36,10 +36,10 @@ GInputStream *gitg_platform_support_http_get_finish (GAsyncResult         *resul
 
 cairo_surface_t *gitg_platform_support_create_cursor_surface (GdkDisplay    *display,
                                                               GdkCursorType  cursor_type,
-                                                              gint          *hot_x,
-                                                              gint          *hot_y,
-                                                              gint          *width,
-                                                              gint          *height);
+                                                              gdouble       *hot_x,
+                                                              gdouble       *hot_y,
+                                                              gdouble       *width,
+                                                              gdouble       *height);
 
 GInputStream *gitg_platform_support_new_input_stream_from_fd (gint     fd,
                                                               gboolean close_fd);
